+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- p-value checking for pandas 3.0 Copy-on-Write compatibility
+
 
 [0.13.3]: https://github.com/pvcaptest/pvcaptest/compare/v0.13.2...v0.13.3
 ## [0.13.3] - 2026-02-06